Get the case report forms

Get the case reports of one or several form(s).

amber.case_report_forms(
  amber,
  study = NULL,
  form = NULL,
  revision = NULL,
  query = list(),
  skip = 0,
  limit = 100,
  df = TRUE
)

Arguments

amber

An Amber object

study

Study identifier (name or id), optional.

form

Form identifier (name or id), optional.

revision

Revision number, optional, default is NULL (means that the case report form uses the latest form revision).

query

The search query

skip

Number of items to skip

limit

Max number of items

df

Return a data.frame (default is TRUE)

See also

Examples

if (FALSE) {
a <- amber.login("https://amber-demo.obiba.org")
amber.case_report_forms(a, form="61e69a22fea2df2f3108b508", skip=0, limit=10)
amber.case_report_forms(a, form="Adult trauma")
amber.case_report_forms(a, study="Trauma Registry", query = list(revision = 1))
amber.case_report_forms(a, query = list(revision = 1))
amber.logout(a)
}